Separate data/ and models/ and extract all model definitions from models/Model.py
