+++ nnlo/examples/example_mnist.py
@@ -1,9 +1,5 @@
 from nnlo.models.model_mnist_tf import make_mnist_model
 
-get_model = make_mnist_model
-#def get_name():
-#    return 'mnist'
-
 def get_all():
     import socket,os,glob
     host = os.environ.get('HOST',os.environ.get('HOSTNAME',socket.gethostname()))
